Все стили объеденины в одном файле. Перенесен функционал преподавателя в новый сервлет. Немного подправлен jsp курсов, для вывода курсов преподавателя. Попытка создания всплывающего текста для подтверждения внесения изменений в имя преподавателя...
