--- controller/src/main/java/by/it/academy/task10/Main.java
@@ -1,7 +1,0 @@
-package by.it.academy.task10;
-
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}
