+++ controller/src/main/webapp/course/newcourse.jsp
@@ -4,8 +4,7 @@
 <head>
     <meta charset="UTF-8"/>
     <title>Academy</title>
-    <link rel="stylesheet" href="../css/style.css">
-    <link rel="stylesheet" href="../css/form.css">
+    <link rel="stylesheet" href="css/style.css">
 </head>
 <body>
 <%@include file="../menu.jsp" %>
