--- controller/src/main/webapp/css/form.css
@@ -1,65 +1,0 @@
-* {
-    box-sizing: border-box;
-}
-
-.name{
-    color: white;
-    font-family: 'Roboto', sans-serif;
-}
-.transparent {
-    position: relative;
-    max-width: 1000px;
-    max-height: 1000px;
-    padding: 60px 50px;
-    margin: 50px auto 0;
-    background: #E6E6FA;
-    background-size: cover;
-}
-.transparent:before {
-    content: "";
-    position: absolute;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    background: linear-gradient(to right bottom, rgba(43, 44, 78, .5), rgba(104, 22, 96, .5));
-}
-.form-inner {
-    position: relative;
-}
-.form-inner h3 {
-    content: "";
-    position: relative;
-    margin-top: 0;
-    color: white;
-    font-family: 'Roboto', sans-serif;
-    font-weight: 300;
-    font-size: 32px;
-    text-transform: uppercase;
-}
-.form-inner h3:after {
-
-    position: absolute;
-    left: 0;
-    bottom: -6px;
-    height: 2px;
-    width: 60px;
-    background: #1762EE;
-}
-.form-inner input {
-    display: block;
-    width: 100%;
-    padding: 0 15px;
-    margin: 20px 0 10px;
-    border-width: 0;
-    line-height: 40px;
-    border-radius: 20px;
-    color: white;
-    background: rgba(255, 255, 255, .2);
-    font-family: 'Roboto', sans-serif;
-}
-
-.form-inner input[type="submit"] {
-    margin-top:50px;
-    background: #2F4F4F;
-}
