+++ controller/src/main/webapp/css/style.css
@@ -1,6 +1,5 @@
-
 body {
-    background: #C0C0C0 url(images/back.jpg) no-repeat fixed;
+    background: #C0C0C0 url("../images/back.jpg") no-repeat fixed;
     -moz-background-size: 100%;
     -webkit-background-size: 100%;
     -o-background-size: 100%;
@@ -8,6 +7,7 @@ body {
     height: 100%;
     margin-left: 0px;
     margin-right: 0px;
+    padding-bottom: 60px;
 }
 
 ul {
@@ -56,4 +56,226 @@ li:hover > ul {
 
 li:hover > ul > li {
     float: none;
+}
+
+table {
+    content: "";
+    position: relative;
+    width: 70%;
+    min-height: 60%;
+    margin-top: 45px;
+    margin-left: auto;
+    margin-right: auto;
+    border: 1px solid #9d82ab;
+    border-collapse: collapse;
+    background-color: #9d82ab;
+}
+
+caption {
+    line-height: 2.0;
+    height: 65px;
+    text-align: left;
+    background-color: #9d82ab;
+    color: #E6E6FA;
+    font-family: 'Roboto', sans-serif;
+    font-weight: 300;
+    font-size: 32px;
+    text-transform: uppercase;
+}
+
+.table th {
+    font-weight: bold;
+    padding: 5px;
+    background-color: #9d82ab;
+    border: 1px solid #9d82ab;
+}
+
+th {
+    color: #E6E6FA;
+    background-color: #2F4F4F;
+    height: 25px;
+}
+
+td {
+    text-align: center;
+}
+
+.table td {
+    border: 1px solid #9d82ab;
+    padding: 5px;
+}
+
+* {
+    box-sizing: border-box;
+}
+
+.name {
+    color: white;
+    font-family: 'Roboto', sans-serif;
+}
+
+.transparent {
+    position: relative;
+    max-width: 1000px;
+    max-height: 1000px;
+    padding: 60px 50px;
+    margin: 50px auto 0;
+    background: #E6E6FA;
+    background-size: cover;
+}
+
+.transparent:before {
+    content: "";
+    position: absolute;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background: linear-gradient(to right bottom, rgba(43, 44, 78, .5), rgba(104, 22, 96, .5));
+}
+
+.form-inner {
+    position: relative;
+}
+
+.form-inner h3 {
+    content: "";
+    position: relative;
+    margin-top: 0;
+    color: white;
+    font-family: 'Roboto', sans-serif;
+    font-weight: 300;
+    font-size: 32px;
+    text-transform: uppercase;
+}
+
+.form-inner h3:after {
+
+    position: absolute;
+    left: 0;
+    bottom: -6px;
+    height: 2px;
+    width: 60px;
+    background: #1762EE;
+}
+
+.form-inner input {
+    display: block;
+    width: 100%;
+    padding: 0 15px;
+    margin: 20px 0 10px;
+    border-width: 0;
+    line-height: 40px;
+    border-radius: 20px;
+    color: white;
+    background: rgba(255, 255, 255, .2);
+    font-family: 'Roboto', sans-serif;
+}
+
+.form-inner input[type="submit"] {
+    margin-top: 50px;
+    background: #2F4F4F;
+}
+
+.foot {
+    position: fixed;
+    left: 0;
+    bottom: 0;
+    width: 100%;
+    height: 50px;
+    padding-left: 0;
+    margin: 0 0 0;
+    background: #2F4F4F;
+}
+
+.textfooter {
+    text-align: center;
+    font-family: arial, serif;
+    color: #fff;
+}
+
+.fade-in-text-sub {
+    vertical-align: middle;
+    color: #590a07;
+    opacity: 0;
+    -webkit-animation: fadeIn linear 3s forwards;
+    -moz-animation: fadeIn linear 3s forwards;
+    -o-animation: fadeIn linear 3s forwards;
+    -ms-animation: fadeIn linear 3s forwards;
+    animation: fadeIn linear 3s forwards;
+    animation-delay: 1s;
+    text-align: center;
+}
+
+#B {
+    vertical-align: middle;
+    color: #87ff01;
+    opacity: 0;
+    -webkit-animation: fadeIn linear 3s forwards;
+    -moz-animation: fadeIn linear 3s forwards;
+    -o-animation: fadeIn linear 3s forwards;
+    -ms-animation: fadeIn linear 3s forwards;
+    animation: fadeIn linear 10s forwards;
+    animation-delay: 1s;
+    text-align: center;
+    position: absolute;
+    top: 200px;
+    left: 200px;
+    width: 100%;
+    font-size: x-large;
+    float: none;
+}
+
+@keyframes fadeIn {
+    0% {
+        opacity: 0;
+    }
+    50% {
+        opacity: 1;
+    }
+    100% {
+        opacity: 0;
+    }
+}
+
+@-moz-keyframes fadeIn {
+    0% {
+        opacity: 0;
+    }
+    100% {
+        opacity: 1;
+    }
+}
+
+@-webkit-keyframes fadeIn {
+    0% {
+        opacity: 0;
+    }
+    100% {
+        opacity: 1;
+    }
+}
+
+@-o-keyframes fadeIn {
+    0% {
+        opacity: 0;
+    }
+    100% {
+        opacity: 1;
+    }
+}
+
+@-ms-keyframes fadeIn {
+    0% {
+        opacity: 0;
+    }
+    100% {
+        opacity: 1;
+    }
+}
+
+@keyframes hide {
+    to {
+        opacity: 0;
+    }
 }
