--- controller/src/main/webapp/css/table.css
@@ -1,45 +1,0 @@
-table {
-    content: "";
-    position: relative;
-    width: 70%;
-    min-height: 60%;
-    margin-top: 45px;
-    margin-left: auto;
-    margin-right: auto;
-    border: 1px solid #9d82ab;
-    border-collapse: collapse;
-    background-color: #9d82ab;
-}
-
-caption {
-    line-height: 2.0;
-    height: 65px;
-    text-align: left;
-    background-color: #9d82ab;
-    color: #E6E6FA;
-    font-family: 'Roboto', sans-serif;
-    font-weight: 300;
-    font-size: 32px;
-    text-transform: uppercase;
-}
-
-.table th {
-    font-weight: bold;
-    padding: 5px;
-    background-color: #9d82ab;
-    border: 1px solid #9d82ab;
-}
-
-th {
-    color:#E6E6FA;
-    background-color: #2F4F4F;
-    height: 25px;
-}
-td{
-    text-align: center;
-}
-
-.table td {
-    border: 1px solid #9d82ab;
-    padding: 5px;
-}
