+++ controller/src/main/webapp/footer.jsp
@@ -6,26 +6,6 @@
   To change this template use File | Settings | File Templates.
 --%>
 <%@page contentType="text/html; charset=UTF-8" language="java" %>
-<style>
-    .foot {
-        position: fixed;
-        left: 0;
-        bottom: 0;
-        width: 100%;
-        height: 80px;
-        margin: 0;
-        padding-left: 0;
-        margin-bottom: 0px;
-        background: #2F4F4F;
-        height: 60px;
-    }
-
-    .textfooter {
-        text-align: center;
-        font-family: arial;
-        color: #fff;
-    }
-</style>
 <footer class="foot">
     <p class = "textfooter">&copy; IT-Academy GroupTask</p>
 </footer>
