--- controller/src/main/webapp/images/stylesheet.css
@@ -1,54 +1,0 @@
-body {
-    background:
-            #C0C0C0
-            url(back.jpg)
-            no-repeat
-            fixed;
-    -moz-background-size: 100%;
-    -webkit-background-size: 100%;
-    -o-background-size: 100%;
-    background-size: 100%;
-}
-ul {
-    list-style: none;
-    margin:0;
-    padding-left: 0;
-    margin-top:25px;
-    background:#2F4F4F;
-    height: 50px;
-}
-a {
-    text-decoration: none;
-    background:#2F4F4F;
-    color:#fff;
-    padding:0px 15px;
-    font-family: arial;
-    line-height:50px;
-    display: block;
-    border-right: 1px solid #677B27;
-    -moz-transition: all 0.3s 0.01s ease;
-    -o-transition: all 0.3s 0.01s ease;
-    -webkit-transition: all 0.3s 0.01s ease;
-}
-a:hover {
-    background:#789c9b;
-}
-li {
-    float:left;
-    position:relative;
-}
-
-
-li > ul {
-    position:absolute;
-    top:25px;
-    display:none;
-}
-
-li:hover > ul {
-    display:block;
-    width:250px;
-}
-li:hover > ul > li {
-    float:none;
-}
