+++ controller/src/main/webapp/mentor/mentors.jsp
@@ -2,33 +2,32 @@
 <%@ page contentType="text/html; charset=UTF-8" language="java" %>
 <%@ page isELIgnored="false"%>
 <!DOCTYPE html>
-<html>
+<html xml:lang="ru">
 <head>
     <meta charset="UTF-8"/>
-    <title>Список преподавателей</title>
-
-    <link rel="stylesheet" href="css/table.css">
+    <title>View mentor list</title>
+    <link rel="stylesheet" href="css/style.css">
 </head>
 <body>
-<%@include file="../menu.jsp" %>
-
+<%@include file="/menu.jsp" %>
+<div id="B"><c:if test="${UpdateSuccessMsg != null}"><c:out value="${UpdateSuccessMsg}"/></c:if></div>
 <table>
     <caption>Список преподавателей</caption>
     <tr>
         <th>Id</th>
         <th>Имя</th>
         <th>Фамилия</th>
-        <th></th>
-        <th></th>
-        <th></th>
+        <th>Курсы</th>
+        <th>Изменить</th>
+        <th>Удалить</th>
     </tr>
             <c:forEach items="${allMentors}" var="mentor">
         <tr>
             <td><c:out value="${mentor.getId()}"/></td>
             <td><c:out value="${mentor.getName()}"/></td>
             <td><c:out value="${mentor.getSurname()}"/></td>
-            <td><a href="mentor/updateMentor.jsp?update_id=${mentor.getId()}&name=${mentor.getName()}&surname=${mentor.getSurname()}">Изменить</a></td>
-            <td><a href="mentor/newmentor.jsp?create_id=${mentor.getId()}&name=${mentor.getName()}&surname=${mentor.getSurname()}">Добавить</a></td>
+            <td><a href="mentors?method=viewcourse&mentor_id=${mentor.getId()}">Курсы</a> </td>
+            <td><a href="mentor/edit.jsp?edit_id=${mentor.getId()}&name=${mentor.getName()}&surname=${mentor.getSurname()}">Изменить</a></td>
             <td><a href="mentor/deleteMentor.jsp?delete_id=${mentor.getId()}">Удалить</a></td>
         </tr>
         </c:forEach>
