+++ controller/src/main/webapp/mentor/viewMentors.jsp
@@ -13,7 +13,7 @@
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="Content-type" content="text/html; charset=UTF-8">
-    <link rel="stylesheet" href="../images/stylesheet.css">
+    <link rel="stylesheet" href="css/style.css">
     <title>View mentor list</title>
 </head>
 <body>
