+++ controller/src/main/webapp/student/deleteStudent.jsp
@@ -6,8 +6,7 @@
 <head>
     <meta charset="UTF-8"/>
     <title>Обновить запись студента</title>
-    <link rel="stylesheet" href="../css/style.css">
-    <link rel="stylesheet" href="../css/form.css">
+    <link rel="stylesheet" href="/css/style.css">
 </head>
 <body>
 <%@include file="../menu.jsp" %>
