--- controller/src/main/webapp/viewMentors.jsp
@@ -1,17 +1,0 @@
-<%--
-  Created by IntelliJ IDEA.
-  User: JustShooter
-  Date: 16.05.2022
-  Time: 20:06
-  To change this template use File | Settings | File Templates.
---%>
-<%@ page contentType="text/html;charset=UTF-8" language="java" %>
-<!DOCTYPE html>
-<html xml:lang="ru">
-<head>
-    <title>View mentor list</title>
-</head>
-<body>
-
-</body>
-</html>
